+++ projects/demo/src/app/app.component.ts
@@ -1,12 +1,21 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { NgxButtonComponent } from 'ngx-zen/button';
-import { NgxAlertComponent } from 'ngx-zen/alert';
+
+import {
+  ZenButtonComponent,
+  ZenAlertComponent,
+  ZenBadgeComponent,
+} from 'ngx-zen/core';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet, NgxButtonComponent, NgxAlertComponent],
+  imports: [
+    RouterOutlet,
+    ZenAlertComponent,
+    ZenBadgeComponent,
+    ZenButtonComponent,
+  ],
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',
 })