+++ projects/demo/tailwind.config.js
@@ -4,7 +4,7 @@ module.exports = {
   theme: {
     extend: {
       fontFamily: {
-        sans: ["Plus Jakarta Sans", "sans-serif"],
+        sans: ["Geist", "serif"],
       },
     },
   },