+++ projects/ngx-zen/core/public-api.ts
@@ -1,0 +1,3 @@
+export * from './button/button.component';
+export * from './alert/alert.component';
+export * from './badge/badge.component';